+++ framework/src/main/java/edu/uci/eecs/wukong/framework/wkpf/WKPFMessageListener.java
@@ -2,25 +2,25 @@ package edu.uci.eecs.wukong.framework.wkpf;
 
 public interface WKPFMessageListener {
 	
-	public void onWKPFGetWuClassList(int sourceId, byte[] message);
+	public void onWKPFGetWuClassList(long sourceId, byte[] message);
 	
-	public void onWKPFGetWuObjectList(int sourceId, byte[] message);
+	public void onWKPFGetWuObjectList(long sourceId, byte[] message);
 	
-	public void onWKPFReadProperty(int sourceId, byte[] message);
+	public void onWKPFReadProperty(long sourceId, byte[] message);
 	
-	public void onWKPFWriteProperty(int sourceId, byte[] message);
+	public void onWKPFWriteProperty(long sourceId, byte[] message);
 	
-	public void onWKPFRequestPropertyInit(int sourceId, byte[] message);
+	public void onWKPFRequestPropertyInit(long sourceId, byte[] message);
 	
-	public void onWKPFMonitoredData(int sourceId, byte[] message);
+	public void onWKPFMonitoredData(long sourceId, byte[] message);
 	
-	public void onWKPFSetLocation(int sourceId, byte[] message);
+	public void onWKPFSetLocation(long sourceId, byte[] message);
 	
-	public void onWKPFGetLocation(int sourceId, byte[] message);
+	public void onWKPFGetLocation(long sourceId, byte[] message);
 	
-	public void onWKPFRemoteProgramOpen(int sourceId, byte[] message);
+	public void onWKPFRemoteProgramOpen(long sourceId, byte[] message);
 	
-	public void onWKPFRemoteProgramWrite(int sourceId, byte[] message);
+	public void onWKPFRemoteProgramWrite(long sourceId, byte[] message);
 	
-	public void onWKPFRemoteProgramCommit(int sourceId, byte[] message);
+	public void onWKPFRemoteProgramCommit(long sourceId, byte[] message);
 }
